Update the position of the click point
When we do find that the element is in a iframe, we need to update the
click position/point so that it takes into account the iframe position
and the click point in the iframe itself.

This new value is returned to caller to use when it comes time to
actually perform a click which happens in the global page context.